Sunon 5015 PWM solution
Sunon EF50151B1-C02C-A99 does not really like PWM control.
Its feature Auto Restart cuts the power at the motor when it is blocked but it also happens with PWM
I wired the fan in paralel with a 22uF capacitor and both in series of a 1Ohm electrical resistence. The capacitor smoothes the PWM signal and it allowes lower speed.
With Kickstart it lets start rotating the fan before applying the PWM signal.